pyproject.toml
[tool.poetry]
name = "globalise-tools"
version = "0.2.0"
description = "A diverse assortment of command-line tools and python libraries for the Globalise Project"
authors = ["Bram Buitendijk <bram.buitendijk@di.huc.knaw.nl>", "Maarten van Gompel <proycon@anaproy.nl>"]
readme = "README.md"
packages = [
    { include = "globalise_tools" },
    { include = "scripts", to = "globalise_tools" },
]

[tool.poetry.dependencies]
annorepo-client = "^0.3.1"
appengine-python-standard = "^1.1.3"
blessedtable = "^1.0.0"
colorama = "^0.4.6"
dataclasses-json = "^0.6.7"
dkpro-cassis = "^0.10.1"
hydra-core = "^1.3.2"
icecream = "^2.1.3"
intervaltree = "^3.1.0"
jsondataclass = "^0.3.0"
pagexml-tools = "^0.7.0"
progressbar2 = "^4.2.0"
pycaprio = "^0.3.0"
python = "^3.12.0"
stam = ">=0.10.0"
tabulate = "^0.9.0"
textrepo-client = "^0.6.0"
uri = { git = "https://github.com/marrow/uri.git", rev = "5b58db87451ca4680004a8993a56bfc4dafff4d4" }
uuid = "^1.30"
xlsxwriter = "^3.0.3"
xmldict = "^0.4.1"
multiprocess = "^0.70.17"

# use `poetry install --with dev-bram` to use these dependencies
loguru = "^0.7.3"
[tool.poetry.group.dev-bram.dependencies]
circuitbreaker = "^2.0.0"
dask = {extras = ["distributed"], version = "^2024.12.1"}
retry = "^0.9.2"
annorepo-client = { path = "/Users/bram/workspaces/annorepo/annorepo-client", develop = true }
pagexml-tools = { path = "/Users/bram/workspaces/pagexml", develop = true }
provenance-client = { path = "/Users/bram/workspaces/provenance/provenance-python-client", develop = true }
textrepo-client = { path = "/Users/bram/workspaces/textrepo/textrepo-client-python", develop = true }

[tool.poetry.scripts]
gt-classify-language = "globalise_tools.scripts.gt_classify_language:main"
gt-extract-lines = "globalise_tools.scripts.gt_extract_lines:main"
gt-xmi-to-wa = "globalise_tools.scripts.gt_xmi_to_wa:main"
gt-fix-reading-order = "globalise_tools.scripts.gt_fix_reading_order:main"
gt-merge-manual-corrections = "globalise_tools.scripts.gt_merge_manual_corrections:main"
gt-align-rgp = "globalise_tools.scripts.gt_align_rgp:main"
gt-align-rgp-lines = "globalise_tools.scripts.gt_align_rgp_lines:main"
version = 'poetry_scripts:version'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
